pyproject.toml
[build-system]
requires = ['setuptools']

[project]
name = 'rl-control-template'
version = '0.0.0'
description = ''
authors = [
    {name = 'Parham Panahi', email = 'parham1@ualberta.ca'},
    {name = 'Andy Patterson', email = 'ap3@ualberta.ca'},
    {name = 'Niko Yasui', email = 'yasuiniko@gmail.com'}
]
requires-python = '>=3.11'
readme = 'README.md'
license = {text = 'MIT'}
classifiers = [
    'Programming Language :: Python :: Implementation :: CPython',
    'Programming Language :: Python :: Implementation :: PyPy',
]
dependencies = [
    # custom minatar with jit compiled environments
    'MinAtar @ git+https://github.com/andnp/MinAtar',
    'foragerenv @ git+https://github.com/andnp/forager',
    "gymnasium[box2d,other]>=1.0.0",
    'optuna',
    'pillow',
    'numba>=0.55.0',
    'chex>=0.1.5',
    'jax>=0.2.14',
    'dm-haiku>=0.0.4',
    'optax>=0.0.8',
    'numpy>=2.0.0',
    'PyFixedReps-andnp>=4.1.2',
    'PyExpUtils-andnp~=8.0',
    'pyrlenvs-andnp~=2.0',
    'tile-coder-rs @ git+https://github.com/andnp/tile-coder-rs.git',
    'ReplayTables-andnp~=8.0',
    'RlGlue-andnp~=2.0',
    'RlEvaluation~=1.0',
    'PyExpPlotting-andnp',
    'matplotlib',
    "ml-instrument>=0.3.1",
    "continual-foragax>=0.12.0",
    "distrax",
    "flashbax>=0.1.3",
    "jax-tqdm>=0.4.0",
    "tol-colors>=2.1.0",
]

[dependency-groups]
cuda = [
    "jax[cuda]>=0.7.1",
]
dev = [
    'pip',
    'ruff',
    'pyright',
    'commitizen',
    'pre-commit',
]

[tool]
[tool.commitizen]
name = 'cz_conventional_commits'
version = '0.0.0'
tag_format = '$version'
version_files = ['pyproject.toml']

[tool.ruff.lint]
select = ['F', 'E', 'W', 'B']
ignore = ['E501', 'E701', 'B023']

[tool.pyright]
include = ['.']
venvPath = '.'
venv = '.venv'
typeCheckingMode = 'standard'
useLibraryCodeForTypes = true
